% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help_functions.R
\name{add_entries_sameuse_onetno_onetimepoint}
\alias{add_entries_sameuse_onetno_onetimepoint}
\title{Function to get sum of  entries  of resource per individual at diff timepoints
if same cateogry has listed multiple time for same id of pariticpant , this
method comes in handy to get the sum}
\usage{
add_entries_sameuse_onetno_onetimepoint(
  use_data,
  timepointcol,
  timepointval,
  idcolumn,
  result_col
)
}
\arguments{
\item{use_data}{the data where the observations are held}

\item{timepointcol}{columnname in the data where the timepoints are noted}

\item{timepointval}{which time point is considered now
at which the descriptive analysis is done}

\item{idcolumn}{id for each participant}

\item{result_col}{name of the column where the sum of entries to be saved}
}
\value{
the data with added sum of resource use
}
\description{
Function to get sum of  entries  of resource per individual at diff timepoints
if same cateogry has listed multiple time for same id of pariticpant , this
method comes in handy to get the sum
}
\examples{
eg_data <- as.data.frame(list(no = c(1, 2, 3, 4),
mark_at_1 = c(12, 7, 23, 45), gender = c("M", "F", "M", "F"),
mark_at_2 = c(12, 34, 89, 45), trialarm = c("1","1","2","2"),
time = c(1,1,2,2), id = c(1, 1, 1, 2)))
add_entries_sameuse_onetno_onetimepoint(eg_data,  "time", 1, "id",
("mark_at_2"))
}
