% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_notused_functions.R
\name{add_probabilities}
\alias{add_probabilities}
\title{Function to add the probabilities in a vector}
\usage{
add_probabilities(probs)
}
\arguments{
\item{probs}{probabilities}
}
\value{
sum of probabilities if success else -1
}
\description{
Function to add the probabilities in a vector
}
\details{
if probabilities not numeric or fall out of range (0 and 1)
return error else add the probabilities together
}
\examples{
add_probabilities(c(0.6, 0.2))
}
