% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adl_scoring.df}
\alias{adl_scoring.df}
\title{adl_scoring table}
\format{
A 41 by 3 dataframe
}
\source{
created on Jan 15, 2020
}
\usage{
adl_scoring.df
}
\description{
adl_scoring table
}
\keyword{datasets}
