% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_model_functions.R
\name{assign_parameters}
\alias{assign_parameters}
\title{Function to assign the values of nested parameters from the parameter list}
\usage{
assign_parameters(param_list)
}
\arguments{
\item{param_list}{list of parameters, some of which can be nested}
}
\value{
list of assigned parameters
}
\description{
Function to assign the values of nested parameters from the parameter list
}
\examples{
param_list = define_parameters(cost_direct_med_A = 1701, cost_comm_care_A = 1055,
cost_direct_med_B = 1774, cost_comm_care_B = 1278, cost_direct_med_C = 6948,
cost_comm_care_C = 2059, cost_zido = 2456, cost_health_A = "cost_direct_med_A+ cost_comm_care_A",
cost_health_B = "cost_direct_med_B+ cost_comm_care_B",
cost_health_C = "cost_direct_med_C + cost_comm_care_C", cost_drug = "cost_zido")
assign_parameters(param_list)
}
