% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_parameter_estimation_functions.R
\name{check_link_glm}
\alias{check_link_glm}
\title{Function to find the keyword for family of distribution in glm}
\usage{
check_link_glm(family, link)
}
\arguments{
\item{family}{family of distribution}

\item{link}{function to be used}
}
\value{
the link if they can be accepted else error
}
\description{
Function to find the keyword for family of distribution in glm
}
\details{
Check and get the link function for the method glm
}
\examples{
check_link_glm("gaussian", "identity")
}
