% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_model_functions.R
\name{check_null_na}
\alias{check_null_na}
\title{Function to check the variable null or NA}
\usage{
check_null_na(variable)
}
\arguments{
\item{variable}{name of variable or list of variable to check}
}
\value{
-1 or -2 as error, else return 0 as success
}
\description{
Function to check the variable null or NA
}
\examples{
var = c("a")
check_null_na(var)
}
