% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_model.R
\name{checks_markov_pick_method}
\alias{checks_markov_pick_method}
\title{Checks the input to run the Markov cycles and picks correct method}
\usage{
checks_markov_pick_method(
  current_strategy,
  initial_state,
  discount,
  method,
  half_cycle_correction,
  startup_cost,
  startup_util
)
}
\arguments{
\item{current_strategy}{strategy object}

\item{initial_state}{value of states initially}

\item{discount}{rate of discount for costs and qalys}

\item{method}{what type of half cycle correction needed}

\item{half_cycle_correction}{boolean to indicate half cycle correction}

\item{startup_cost}{cost of states initially}

\item{startup_util}{utility of states initially if any}
}
\value{
changed method name
}
\description{
Checks the input to run the Markov cycles and picks correct method
}
\examples{
tmat <- rbind(c(1, 2), c(3, 4))
colnames(tmat) <- rownames(tmat) <- c("Healthy", "Dead")
tm <- populate_transition_matrix(2, tmat, c(0.5, 0.5, 0, 1))
a <- health_state("Healthy", 1, 1, 0, FALSE)
b <- health_state("Dead", 1, 0, 0, TRUE)
health_states <- combine_state(a, b)
this.strategy <- strategy(tm, health_states, "intervention")
checks_markov_pick_method(this.strategy, c(1, 0), c(0, 0),
"half cycle correction", TRUE,NULL,NULL)
}
