% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_help_functions.R
\name{checksum_rowprob}
\alias{checksum_rowprob}
\title{Function to check the sum of probabilities in a vector}
\usage{
checksum_rowprob(avector)
}
\arguments{
\item{avector}{vector of probabilities}
}
\value{
sum of probabilities if success else error
}
\description{
Function to check the sum of probabilities in a vector
}
\examples{
checksum_rowprob(c(0.6,0.4))
}
