% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_cost_analysis_functions.R
\name{convert_freq_diff_basis}
\alias{convert_freq_diff_basis}
\title{Convert frequency medication to given basis}
\usage{
convert_freq_diff_basis(freq_given, basis = "day")
}
\arguments{
\item{freq_given}{given frequency}

\item{basis}{given basis, default is daily}
}
\value{
converted frequency
}
\description{
Convert frequency medication to given basis
}
\examples{
convert_freq_diff_basis("once daily")
convert_freq_diff_basis("bd", "week")
convert_freq_diff_basis("Every 4 days", "day")
}
