% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_help_functions.R
\name{convert_time_unit}
\alias{convert_time_unit}
\title{Convert unit of time to another}
\usage{
convert_time_unit(given_time, basis = "day")
}
\arguments{
\item{given_time}{given unit of time}

\item{basis}{given basis, default is "day"}
}
\value{
converted unit
}
\description{
Convert unit of time to another
}
\examples{
convert_time_unit("day")
convert_time_unit("sec")
convert_time_unit("sec", "minute")
}
