% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_cost_analysis_functions.R
\name{convert_to_given_timeperiod}
\alias{convert_to_given_timeperiod}
\title{Convert period to given basis}
\usage{
convert_to_given_timeperiod(given_time, basis_time = "day")
}
\arguments{
\item{given_time}{given time}

\item{basis_time}{given basis, default is "day"}
}
\value{
converted unit
}
\description{
Convert period to given basis
}
\examples{
convert_to_given_timeperiod("4 weeks")
convert_to_given_timeperiod("a month")
convert_to_given_timeperiod("1 week")
}
