% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_help_functions.R
\name{convert_unit_diff_basis}
\alias{convert_unit_diff_basis}
\title{Convert unit strength to given basis}
\usage{
convert_unit_diff_basis(given_unit, basis = "mg")
}
\arguments{
\item{given_unit}{given unit}

\item{basis}{given basis, default is "mg"}
}
\value{
converted unit
}
\description{
Convert unit strength to given basis
}
\examples{
convert_unit_diff_basis("mg")
convert_unit_diff_basis("mg", "mcg")
convert_unit_diff_basis("l","ml")
}
