% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_help_functions.R
\name{convert_volpertimediff_basis}
\alias{convert_volpertimediff_basis}
\title{Convert volume per time  to given basis}
\usage{
convert_volpertimediff_basis(given_unit, basis = "ml/hour")
}
\arguments{
\item{given_unit}{given unit}

\item{basis}{given basis, default is "mg"}
}
\value{
converted unit
}
\description{
Convert volume per time  to given basis
}
\examples{
convert_volpertimediff_basis("ml/day")
convert_volpertimediff_basis("mcl/day")
convert_volpertimediff_basis("ml/hour")
}
