% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_cost_analysis_functions.R
\name{convert_weight_diff_basis}
\alias{convert_weight_diff_basis}
\title{Convert unit strength to given basis}
\usage{
convert_weight_diff_basis(given_unit, basis = "mg")
}
\arguments{
\item{given_unit}{given unit}

\item{basis}{given basis, default is "mg"}
}
\value{
converted unit
}
\description{
Convert unit strength to given basis
}
\examples{
convert_weight_diff_basis("mg")
convert_weight_diff_basis("kilogram", "micro gram")
}
