% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_cost_analysis_functions.R
\name{convert_wtpertimediff_basis}
\alias{convert_wtpertimediff_basis}
\title{Convert weight per time  to given basis}
\usage{
convert_wtpertimediff_basis(given_unit, basis = "mcg/hour")
}
\arguments{
\item{given_unit}{given unit}

\item{basis}{given basis, default is "mg"}
}
\value{
converted unit
}
\description{
Convert weight per time  to given basis
}
\examples{
convert_wtpertimediff_basis("mg/day")
convert_wtpertimediff_basis("mcg/day")
convert_wtpertimediff_basis("mg/hour")
}
