% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_cost_analysis_functions.R
\name{convert_wtpervoldiff_basis}
\alias{convert_wtpervoldiff_basis}
\title{Convert wt per unit volume  to given basis}
\usage{
convert_wtpervoldiff_basis(given_unit, basis = "mg/ml")
}
\arguments{
\item{given_unit}{given unit}

\item{basis}{given basis, default is "mg/ml"}
}
\value{
converted unit
}
\description{
Convert wt per unit volume  to given basis
}
\examples{
convert_wtpervoldiff_basis("g/ml")
}
