% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cost_data}
\alias{cost_data}
\title{cost matrix}
\format{
A 11 by 2 dataframe
}
\source{
created on Nov 26, 2019 from
tmat <- rbind(c(1, 2), c(3, 4))
colnames(tmat) <- rownames(tmat) <- c("Healthy", "Dead")
tm <- transition_matrix(2, tmat, c(0.5, 0.5, 0, 1))
a <- health_state("Healthy", 1, 1, FALSE)
b <- health_state("Dead", 1, 0, TRUE)
health_states <- combine_state(a, b)
this.strategy <- strategy(tm, health_states, "intervention")
}
\usage{
cost_data
}
\description{
cost matrix
}
\keyword{datasets}
