% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_parameter_estimation_survival.R
\name{create_new_dataset}
\alias{create_new_dataset}
\title{create new dataset while keeping
cox regression results and returned coefficients}
\usage{
create_new_dataset(var, covar, dataset, categorical)
}
\arguments{
\item{var}{variable for which the levels have to be identified
usually indep variable}

\item{covar}{the other covariates}

\item{dataset}{the dataset where these variables contain}

\item{categorical}{are these variables categorical? True of false}
}
\value{
new data frame
}
\description{
create new dataset while keeping
cox regression results and returned coefficients
}
\examples{
dataset <- survival::lung
new = create_new_dataset("status", c("age"), dataset, c(FALSE))
}
