% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_help_functions.R
\name{define_parameters}
\alias{define_parameters}
\title{Function to return a list of parameters given}
\usage{
define_parameters(...)
}
\arguments{
\item{...}{any parameters set of name value pairs expected}
}
\value{
a list of parameters
}
\description{
Function to return a list of parameters given
}
\examples{
define_parameters(rr=1)
}
