% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_model_functions.R
\name{define_transition_table}
\alias{define_transition_table}
\title{Define the table for transition}
\usage{
define_transition_table(tmat)
}
\arguments{
\item{tmat}{transition matrix in the format as in package 'mstate'}
}
\value{
the transition table with the probabilities
}
\description{
Define the table for transition
}
\examples{
tmat <- rbind(c(1, 2), c(3, 4))
colnames(tmat) <- rownames(tmat) <- c("Healthy", "Dead")
define_transition_table(tmat)
}
