% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_cost_analysis_functions.R
\name{encode_codes_data}
\alias{encode_codes_data}
\title{Function to get the codes and the corresponding entries}
\usage{
encode_codes_data(list_code_values, data_column_nos, the_data)
}
\arguments{
\item{list_code_values}{list of codes and values, given as list of lists}

\item{data_column_nos}{the column numbers of data to look for the entries}

\item{the_data}{the data where to look for}
}
\value{
weight and vol units
}
\description{
Function to get the codes and the corresponding entries
}
\examples{
data_file <- system.file("extdata", "medication_liq_codes.xlsx",
package = "packDAMipd")
ind_part_data <- load_trial_data(data_file)
data_column_nos = c(2,12)
list_of_code_names = list(c(1, 2),c("Morphine", "Oxycodone"))
encode_codes_data(list_of_code_names, data_column_nos, ind_part_data)
}
