% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_model_functions.R
\name{eval_assign_trans_prob}
\alias{eval_assign_trans_prob}
\title{Attribute parameters to probabilities of transition matrix}
\usage{
eval_assign_trans_prob(tm, parameter_values)
}
\arguments{
\item{tm}{A transition matrix in the format from the package 'mstate'}

\item{parameter_values}{name value pairs of parameter values in the probability matrix}
}
\value{
the transition table with the probabilites
}
\description{
Attribute parameters to probabilities of transition matrix
}
\examples{
tmat <- rbind(c(1, 2), c(3, 4))
colnames(tmat) <- rownames(tmat) <- c("Healthy", "Dead")
tmat<-transition_matrix(2, tmat, list_prob = c("p1", "p2", "p3", "p4"))
tmat_assigned<-eval_assign_trans_prob(tmat,c(p1=0.2, p2=0.3, p3=0.4, p4=0.5))
}
