% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_model_functions.R
\name{eval_assign_values_states}
\alias{eval_assign_values_states}
\title{Attribute values in health states}
\usage{
eval_assign_values_states(health_states, assigned_param)
}
\arguments{
\item{health_states}{list of health_state objects}

\item{assigned_param}{name value pairs of parameter values in the probability matrix}
}
\value{
health states with assigned values
}
\description{
Attribute values in health states
}
\examples{
well <-  health_state("well", cost ="cost_A", utility = 1)
disabled <- health_state("disabled", cost = 100, utility = 1)
dead <- health_state("dead", cost = 0, utility = 0)
tmat <- rbind(c(1, 2,3), c(NA, 4,5), c(NA,NA,6))
colnames(tmat) <- rownames(tmat) <- c("well","disabled" ,"dead")
tm <- transition_matrix(3, tmat, c(0.6, 0.2, 0.2, 0.6, 0.4,1))
health_states <- combine_state(well,disabled,dead)
eval_assign_values_states(health_states, c(cost_A= 100))
}
