% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_help_functions.R
\name{find_keyword_regression_method}
\alias{find_keyword_regression_method}
\title{Function to find the keyword for regression methods}
\usage{
find_keyword_regression_method(text, additional_info = NA)
}
\arguments{
\item{text}{regresssion method}

\item{additional_info}{additional infomration required}
}
\value{
the keyword that should be used in R for regression analysis
}
\description{
Function to find the keyword for regression methods
}
\examples{
find_keyword_regression_method("linear")
}
