% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_help_functions.R
\name{find_param_from_def_prob_distrbn}
\alias{find_param_from_def_prob_distrbn}
\title{Function to return parameters from a given expression separated by comma,}
\usage{
find_param_from_def_prob_distrbn(expr)
}
\arguments{
\item{expr}{an expression}
}
\value{
parameters in the expression expr
}
\description{
Function to return parameters from a given expression separated by comma,
}
\examples{
find_param_from_def_prob_distrbn("gamma(mean = 10, sd =1)")
}
