% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_help_functions.R
\name{find_parameters_btn_operators}
\alias{find_parameters_btn_operators}
\title{Function to return parameters with in a expression containing operators}
\usage{
find_parameters_btn_operators(expr)
}
\arguments{
\item{expr}{an expression}
}
\value{
parameters in the expression expr
}
\description{
Function to return parameters with in a expression containing operators
}
\examples{
find_parameters_btn_operators("a+b")
}
