% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_help_functions.R
\name{find_required_parameter_combs}
\alias{find_required_parameter_combs}
\title{Function to find the parameters that determine the given probability distribution}
\usage{
find_required_parameter_combs(name_distri)
}
\arguments{
\item{name_distri}{name of the probability distribution}
}
\value{
the parameters that determine the distribution
}
\description{
Function to find the parameters that determine the given probability distribution
}
\examples{
find_required_parameter_combs("gamma")
}
