% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help_functions.R
\name{find_rowwise_sum_multiplecol}
\alias{find_rowwise_sum_multiplecol}
\title{Function to get sum of  multiple columns of observations}
\usage{
find_rowwise_sum_multiplecol(the_data, colnames, sumcolname)
}
\arguments{
\item{the_data}{the data where the observations are held}

\item{colnames}{columnname in the data whose sum to be obtianed}

\item{sumcolname}{name of the new column where sum to be saved}
}
\value{
the data with added sum
}
\description{
Function to get sum of  multiple columns of observations
}
\examples{
eg_data <- as.data.frame(list(no = c(1, 2, 3, 4),
mark_at_1 = c(12, 7, 23, 45), gender = c("M", "F", "M", "F"),
mark_at_2 = c(12, 34, 89, 45), trialarm = c("1","1","2","2"),
time = c(1,1,2,2), id = c(1, 1, 1, 2)))
find_rowwise_sum_multiplecol(eg_data, c("mark_at_1","mark_at_2"),
"totalmarks")
}
