% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_parameter_estimation_functions.R
\name{find_survreg_distribution}
\alias{find_survreg_distribution}
\title{Function to find the keyword for survreg distribution}
\usage{
find_survreg_distribution(text)
}
\arguments{
\item{text}{distribution}
}
\value{
the keyword - the name of distribution
}
\description{
Function to find the keyword for survreg distribution
}
\details{
For surveg method, find the distribution
}
\examples{
find_survreg_distribution("weibull")
}
