% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_help_functions.R
\name{form_expression_lm}
\alias{form_expression_lm}
\title{Form expression to use with lm()}
\usage{
form_expression_lm(param_to_be_estimated, indep_var, covariates,
  interaction)
}
\arguments{
\item{param_to_be_estimated}{parameter of interest}

\item{indep_var}{the independent variable (column name in data file)}

\item{covariates}{list of covariates - calculations to be done before passing}

\item{interaction}{boolean value to indicate interaction in the case of linear regression,
false by default}
}
\value{
the results of the regression analysis
}
\description{
Form expression to use with lm()
}
\examples{
formula = form_expression_lm("gre", dataset = mydata, indep_var = "gpa", covariates = NA, interaction = FALSE)
}
