% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_help_functions.R
\name{form_expression_mixed_model}
\alias{form_expression_mixed_model}
\title{Form expression to use with mixed models}
\usage{
form_expression_mixed_model(param_to_be_estimated, dataset, fix_eff,
  fix_eff_interact_vars, random_intercept_vars,
  nested_intercept_vars_pairs, cross_intercept_vars,
  uncorrel_slope_intercept_pairs, random_slope_intercept_pairs, family,
  link)
}
\arguments{
\item{param_to_be_estimated}{column name of dependent variable}

\item{dataset}{a dataframe}

\item{fix_eff}{names of variables as fixed effect predictors}

\item{fix_eff_interact_vars, }{if interaction -true}

\item{random_intercept_vars, }{names of variables for random intercept}

\item{nested_intercept_vars_pairs, }{those of the random intercept variables with nested effect}

\item{cross_intercept_vars, }{those of the random intercept variables with crossed effect}

\item{uncorrel_slope_intercept_pairs, }{variables with correlated intercepts}

\item{random_slope_intercept_pairs, }{random slopes intercept pairs - this is alist of paired variables}

\item{family, }{family of distribution for non gaussian distribution of predicted variable}

\item{link, }{link function for the variance}
}
\value{
result regression result with plot if success and -1, if failure
}
\description{
Form expression to use with mixed models
}
\details{
Form the expression for mixed model
}
\examples{
dt <- read.table("http://bayes.acs.unt.edu:8083/BayesContent/class/Jon/R_SC/Module9/lmm.data.txt",
  header = TRUE, sep = ",", na.strings = "NA", dec = ".", strip.white = TRUE
)
formula <- form_expression_mixed_model("extro",
  dataset = dataset,
  fix_eff = c("open", "agree", "social"),
  fix_eff_interact_vars = NULL,
  random_intercept_vars = c("school", "class"),
  nested_intercept_vars_pairs = list(c("school", "class")),
  cross_intercept_vars = NULL,
  uncorrel_slope_intercept_pairs = NULL,
  random_slope_intercept_pairs = NULL, family = "binomial", link = NA
)
}
