% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_parameter_estimation_functions.R
\name{form_expression_mixed_model_lme4}
\alias{form_expression_mixed_model_lme4}
\title{Form expression to use with mixed models}
\usage{
form_expression_mixed_model_lme4(
  param_to_be_estimated,
  dataset,
  fix_eff,
  fix_eff_interact_vars,
  random_intercept_vars,
  nested_intercept_vars_pairs,
  cross_intercept_vars_pairs,
  uncorrel_slope_intercept_pairs,
  random_slope_intercept_pairs,
  family,
  link
)
}
\arguments{
\item{param_to_be_estimated}{column name of dependent variable}

\item{dataset}{a dataframe}

\item{fix_eff}{names of variables as fixed effect predictors}

\item{fix_eff_interact_vars, }{if interaction -true}

\item{random_intercept_vars, }{names of variables for random intercept}

\item{nested_intercept_vars_pairs, }{those of the random intercept variables
with nested effect}

\item{cross_intercept_vars_pairs, }{those of the random intercept variables
with crossed effect}

\item{uncorrel_slope_intercept_pairs, }{variables with correlated intercepts}

\item{random_slope_intercept_pairs, }{random slopes intercept pairs -
this is a list of paired variables}

\item{family, }{family of distribution for non gaussian distribution of
predicted variable}

\item{link, }{link function for the variance}
}
\value{
result regression result with plot if success and -1, if failure
}
\description{
Form expression to use with mixed models
}
\details{
Form the expression for mixed model
}
\examples{
\donttest{
datafile <- system.file("extdata", "data_linear_mixed_model.csv",
package = "packDAMipd")
dt = utils::read.csv(datafile, header = TRUE)
formula <- form_expression_mixed_model_lme4("extro",
  dataset = dt,
  fix_eff = c("open", "agree", "social"),
  fix_eff_interact_vars = NULL,
  random_intercept_vars = c("school", "class"),
  nested_intercept_vars_pairs = list(c("school", "class")),
  cross_intercept_vars_pairs = NULL,
  uncorrel_slope_intercept_pairs = NULL,
  random_slope_intercept_pairs = NULL, family = "binomial", link = NA
)
}
}
