% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_cost_analysis_functions.R
\name{generate_wt_time_units}
\alias{generate_wt_time_units}
\title{Function to get the weight and time units}
\usage{
generate_wt_time_units()
}
\value{
weight and time units
}
\description{
Function to get the weight and time units
}
\examples{
ans <- generate_wt_time_units()
}
