% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_cost_analysis_functions.R
\name{generate_wt_vol_units}
\alias{generate_wt_vol_units}
\title{Function to get the weight and volume units}
\usage{
generate_wt_vol_units()
}
\value{
weight and vol units
}
\description{
Function to get the weight and volume units
}
\examples{
ans <- generate_wt_vol_units()
}
