% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3a_trialdata_analysis_input_functions.R
\name{get_colnames_codedvalues}
\alias{get_colnames_codedvalues}
\title{Function to keep the column name, coded values and non response code
into a dataframe}
\usage{
get_colnames_codedvalues(variable, name, code, nrcode = NA)
}
\arguments{
\item{variable, }{name of the variable in the column}

\item{name, }{column name}

\item{code, }{coded values}

\item{nrcode, }{code for non response}
}
\value{
data frame with all the above information
}
\description{
Function to keep the column name, coded values and non response code
into a dataframe
}
\examples{
get_colnames_codedvalues("arm", "pat_trial_arm", c("Y", "N"))
}
