% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3c_costing_resource_use_functions.R
\name{get_cost_AandE_description}
\alias{get_cost_AandE_description}
\title{Function to extract the unit cost by descirption of AandE att
matching description}
\usage{
get_cost_AandE_description(
  description,
  type_admit,
  ref_cost_data_file,
  col_name_description,
  unit_cost_col,
  type_admit_col,
  sheet = NULL
)
}
\arguments{
\item{description}{description of the AE attendance}

\item{type_admit}{term indicating admission and type of attendance}

\item{ref_cost_data_file}{file that has unit cost}

\item{col_name_description}{name of the column that has the description}

\item{unit_cost_col}{name of the column with the unit cost}

\item{type_admit_col}{colname that descirbes type of the attendance and}

\item{sheet}{sheet if excel file is given}
}
\value{
unit cost the unit cost matching the hrg code
}
\description{
Function to extract the unit cost by descirption of AandE att
matching description
}
\examples{
ref_cost_data_file <- system.file("extdata",
"National_schedule_of_NHS_costs_2019_AandE.csv", package = "packDAMipd")
re = get_cost_AandE_description("Emergency Medicine", "T01A",
ref_cost_data_file, "Currency_Description", "National_Average_Unit_Cost",
"Service_Code")
}
