% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_cost_analysis_functions.R
\name{get_doses_combination}
\alias{get_doses_combination}
\title{Convert the combined dose to its individual component numerical value
or can be unit/unit}
\usage{
get_doses_combination(the_string, separator = "/")
}
\arguments{
\item{the_string}{given combined unit}

\item{separator}{given character for separation , default is "/"}
}
\value{
separated texts
}
\description{
Convert the combined dose to its individual component numerical value
or can be unit/unit
}
\examples{
get_doses_combination("g/ml")
}
