% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_cost_analysis_functions.R
\name{get_doses_combination_units}
\alias{get_doses_combination_units}
\title{Convert the combined dose to its individual component numerical value
and units}
\usage{
get_doses_combination_units(the_string, separator = "/")
}
\arguments{
\item{the_string}{given combined unit}

\item{separator}{given character for separation , default is "/"}
}
\value{
separated numerical value and its units
}
\description{
Convert the combined dose to its individual component numerical value
and units
}
\examples{
get_doses_combination_units("10g/2ml")
}
