% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3a_trialdata_analysis_input_functions.R
\name{get_eq5d_details}
\alias{get_eq5d_details}
\title{Function to get the details of the EQ5D column}
\usage{
get_eq5d_details(trialdata)
}
\arguments{
\item{trialdata, }{data containing individual level trial data}
}
\value{
the name of the variable related to EQ5D and the unique contents
if success, else error
}
\description{
Function to get the details of the EQ5D column
}
\details{
Specific to the EQ5D data - the column names are given as certain sets,
Tried to give 15 sets as the column names
}
\examples{
get_eq5d_details(data.frame(
  "MO" = c(1, 2), "SC" = c(1, 2), "UA" = c(1, 2),
  "PD" = c(1, 2), "AD" = c(1, 2)
))
}
