% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_help_functions.R
\name{get_extension}
\alias{get_extension}
\title{Function to get extension of a filename}
\usage{
get_extension(file)
}
\arguments{
\item{file}{name of a file}
}
\value{
the extension
}
\description{
Function to get extension of a filename
}
\examples{
get_extension("data.txt")
}
