% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3b_outcome_analysis_functions.R
\name{get_mean_sd_age}
\alias{get_mean_sd_age}
\title{Function to return mean age from a data frame}
\usage{
get_mean_sd_age(this_data, age_nrcode)
}
\arguments{
\item{this_data}{the data containing column with age}

\item{age_nrcode}{non response code}
}
\value{
mean and sd, if success -1, if failure
}
\description{
Function to return mean age from a data frame
}
\details{
Age data is complete with the nr code given and get the mean and sd
}
\examples{
this_data <- as.data.frame(cbind(num = c(1, 2, 3, 4),
age = c(14, 25, 26, 30)))
get_mean_sd_age(this_data, NA)
}
