% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_parameter_estimation_functions.R
\name{get_name_value_probdistrb_def}
\alias{get_name_value_probdistrb_def}
\title{Function to return the two parameters from a given expression separated
by comma,}
\usage{
get_name_value_probdistrb_def(expr)
}
\arguments{
\item{expr}{an expression}
}
\value{
parameters in the expression expr
}
\description{
Function to return the two parameters from a given expression separated
by comma,
}
\details{
It will return the parameters of the distribution separated by commas
and given in usual notation as brackets. It will identify those in between
first occurrence of "( "and last occurrence of ")"
and from the characters in between search for comma to indicate different
parameters then it will extract (from those extracted parameters separated
by commas) that on the left side of "equal" sign
get_name_value_probdistrb_def("gamma(mean = sqrt(2), b =17)") will be ok
but get_name_value_probdistrb_def("gamma(shape, scale")) and
get_name_value_probdistrb_def("gamma(shape =1 & scale =1")) will show error
}
\examples{
get_name_value_probdistrb_def("gamma(mean = 10, sd =1)")
}
