% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3a_trialdata_analysis_input_functions.R
\name{get_outcome_details}
\alias{get_outcome_details}
\title{Function to get the details of the outcome column}
\usage{
get_outcome_details(trialdata, name, related_words, multiple = FALSE)
}
\arguments{
\item{trialdata, }{data containing individual level trial data}

\item{name, }{name of the variable}

\item{related_words, }{probable column names}

\item{multiple, }{indicates true if there are multiple columns}
}
\value{
the name of the variable related to health outcome (any) and
the unique contents if success, else error
}
\description{
Function to get the details of the outcome column
}
\details{
if the words related to outcome is given, the function will get the
columns and the codes used for the outcome, the difference here is that
certain outcomes can  be distributed in multiple columns
}
\examples{
get_outcome_details(
  data.frame("qol.MO" = c(1, 2), "qol.PD" = c(1, 2), "qol.AD" = c(1, 2)),
  "eq5d", "qol", TRUE
)
}
