% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_parameter_estimation_functions.R
\name{get_parameter_def_distribution}
\alias{get_parameter_def_distribution}
\title{Get the definition of given parameter distribution defined in a file}
\usage{
get_parameter_def_distribution(parameter, paramfile, strategycol = NA,
  strategyname = NA)
}
\arguments{
\item{parameter}{parameter of interest}

\item{paramfile}{data file to be provided}

\item{strategycol}{treatment strategy column name}

\item{strategyname}{treatment strategy name in the column strategycol}
}
\value{
the definition of parameter from the given distribution
}
\description{
the file should have these column names (atleast)
Parameter,	Description,	Strategy,	Value,	Distribution,	Param1_name, Param1_value,
Param2_name,	Param2_value
}
\examples{
a <- get_parameter_def_distribution("rr", paramfile=
system.file("extdata", "table_param.csv", package = "MarkovModel"))
}
