% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_parameter_estimation_functions.R
\name{get_parameter_direct}
\alias{get_parameter_direct}
\title{Get the parameter values from reading a file}
\usage{
get_parameter_direct(parameter, param_value)
}
\arguments{
\item{parameter}{parameter of interest}

\item{param_value}{parameter value to be assigned}
}
\value{
the paramvalue
}
\description{
Get the parameter values from reading a file
}
\examples{
a <- get_parameter_direct("cost_IT", param_value= 100)
}
