% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_cost_analysis_functions.R
\name{get_single_col_multiple_pattern}
\alias{get_single_col_multiple_pattern}
\title{Function to get cols for the pattern given}
\usage{
get_single_col_multiple_pattern(pattern, the_data)
}
\arguments{
\item{pattern}{the pattern to look for}

\item{the_data}{data where to look at}
}
\value{
zero or -1
}
\description{
Function to get cols for the pattern given
}
\examples{
the_data <- as.data.frame(cbind(c("one", "two"), c("a", "b"), c("aa", "bb")))
colnames(the_data) <- c("name", "brand_one", "two")
get_single_col_multiple_pattern(c("brand", "trade"), the_data)
}
