% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_parameter_estimation_functions.R
\name{get_slope_intercept}
\alias{get_slope_intercept}
\title{help function to keep slope and intercept portion ready in mixed model
expression}
\usage{
get_slope_intercept(
  expression,
  random_intercept_vars,
  random_slope_intercept_pairs,
  uncorrel_slope_intercept_pairs
)
}
\arguments{
\item{expression}{expression created so far}

\item{random_intercept_vars, }{names of variables for random intercept}

\item{random_slope_intercept_pairs, }{random slopes intercept pairs
this is a list of paired variables}

\item{uncorrel_slope_intercept_pairs, }{variables with correlated intercepts}
}
\value{
expression expression created
}
\description{
help function to keep slope and intercept portion ready in mixed model
expression
}
