% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3a_trialdata_analysis_input_functions.R
\name{get_timepoint_details}
\alias{get_timepoint_details}
\title{Function to get the details of the time point column}
\usage{
get_timepoint_details(trialdata)
}
\arguments{
\item{trialdata, }{data containing individual level trial data}
}
\value{
the name of the variable related to time point and
the unique contents if success, else error
}
\description{
Function to get the details of the time point column
}
\details{
expecting the data contains the information on timepoints
preferably  column names "time point", "times" or "time" or "timepoint".
If multiple column names match these, then first match will be chosen.
}
\examples{
get_timepoint_details(data.frame("time" = c(21, 15),
"arm" = c("control", "intervention")))
}
