% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3a_trialdata_analysis_input_functions.R
\name{get_trial_arm_details}
\alias{get_trial_arm_details}
\title{Function to get the details of the trial arm}
\usage{
get_trial_arm_details(trialdata)
}
\arguments{
\item{trialdata, }{data containing individual level trial data}
}
\value{
the name of the variable related to trial arm and the unique contents
if success, else error
}
\description{
Function to get the details of the trial arm
}
\details{
expecting the data contains the information on trial arm
preferably  column names "arm", "trial" or "trial arm". If
multiple column names
match these, then first match will be chosen.
}
\examples{
get_trial_arm_details(data.frame(
  "Age" = c(21, 15),
  "arm" = c("control", "intervention")
))
}
