% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_model_functions.R
\name{get_var_state}
\alias{get_var_state}
\title{Get the attribute for the health state}
\usage{
get_var_state(state, var)
}
\arguments{
\item{state}{object of class health state}

\item{var}{attribute of the health state}
}
\value{
modified health state
}
\description{
Get the attribute for the health state
}
\details{
After checking the given state is a health state and given variable
is defined in the health state, the value of the variable is returned
}
\examples{
get_var_state(health_state("IT", 100, 0.4, 0, FALSE), "cost")
}
