% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_model_functions.R
\name{init_trace}
\alias{init_trace}
\title{Define an all zero trace matrix}
\usage{
init_trace(health_states, cycles)
}
\arguments{
\item{health_states}{health states}

\item{cycles}{no of cycles}
}
\value{
trace matrix -all zero
}
\description{
Define an all zero trace matrix
}
\details{
Initialise the trace matrix with all zeros
trace matrix will be with no_cycles+1 by no_states matrix
}
\examples{
a <- health_state("Healthy", 1, 1, 0, FALSE)
b <- health_state("Dead", 1, 0.5, 0, FALSE)
health_states <- combine_state(a, b)
init_trace(health_states, 10)
}
