% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/4a_deterministic_sensitivity_analysis_functions.R
\name{keep_results_plot_dsa}
\alias{keep_results_plot_dsa}
\title{Function to do some checks before plotting sensitivity analysis results}
\usage{
keep_results_plot_dsa(
  result_dsa_control,
  plotfor,
  result_dsa_treat,
  plot_variable,
  threshold,
  comparator
)
}
\arguments{
\item{result_dsa_control}{result from deterministic sensitivity analysis
for first or control model}

\item{plotfor}{the variable to plotfor e.g. cost, utility NMB etc}

\item{result_dsa_treat}{result from deterministic sensitivity analysis
for the comparative Markov model}

\item{plot_variable}{variable for plotting}

\item{threshold}{threshold value of WTP}

\item{comparator}{the strategy to be compared with}
}
\value{
results to plot dsa
}
\description{
Function to do some checks before plotting sensitivity analysis results
}
