% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_parameter_estimation_functions.R
\name{make_string_covariates}
\alias{make_string_covariates}
\title{Make a string of covariates using addition sign}
\usage{
make_string_covariates(covariates)
}
\arguments{
\item{covariates}{covariates}
}
\value{
expression of covariates
}
\description{
Make a string of covariates using addition sign
}
\examples{
make_string_covariates(c("open", "grade"))
}
\keyword{internal}
