% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help_functions.R
\name{map_resource_use_categories}
\alias{map_resource_use_categories}
\title{Function to to read the text form of resource use and replace it with
standard texts of resoure use ie. some one can describe GP visit as GP surgery
visit, surgery visit or general practioners visit etc. Here all these texts
should be given in a excel or csv file and then corresponidng standard form
will be read from the file and will be replaced.}
\usage{
map_resource_use_categories(
  the_data,
  service_actual,
  new_column,
  mapped_data,
  mapped_use,
  analysis,
  replace_only,
  relevant_column = NULL,
  check_value_relevant = NULL,
  nhs_use_column = NULL,
  check_value_nhs_use = NULL
)
}
\arguments{
\item{the_data}{the data where the observations are held}

\item{service_actual}{columna name of the actual service use}

\item{new_column}{the name of the column where the mapped resource use to be}

\item{mapped_data}{data where the service name and mapped service name
has been stored}

\item{mapped_use}{columan name of mapped resource use in mapped_data}

\item{analysis}{base case or secondary}

\item{replace_only}{if we want to replace only certain resource use}

\item{relevant_column}{the name of the column where the mapped resource use
is indicated as relevant or not}

\item{check_value_relevant}{how is the mapped resource
is indicated as relevant by a value}

\item{nhs_use_column}{the name of the column where the mapped resource use
comes under NHS or not}

\item{check_value_nhs_use}{value that is used to indicated the nhs use}
}
\value{
the data with added sum
}
\description{
Function to to read the text form of resource use and replace it with
standard texts of resoure use ie. some one can describe GP visit as GP surgery
visit, surgery visit or general practioners visit etc. Here all these texts
should be given in a excel or csv file and then corresponidng standard form
will be read from the file and will be replaced.
}
