% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_model_functions.R
\name{markov_model}
\alias{markov_model}
\title{Definition of Markov model and trace}
\usage{
markov_model(this_strat, cycles, initial_state, overhead_costs,
  discount = c(0, 0), parameter_values = NULL)
}
\arguments{
\item{this_strat}{strategy object}

\item{cycles}{no of cycles}

\item{initial_state}{value of states initially}

\item{overhead_costs}{any overhead costs for being in each state}

\item{discount}{rate of discount for costs and qalys}

\item{parameter_values}{parameter values for assigning health states and probabilities}
}
\value{
Markov trace
}
\description{
Definition of Markov model and trace
}
\examples{
tmat <- rbind(c(1, 2), c(3, 4))
colnames(tmat) <- rownames(tmat) <- c("Healthy", "Dead")
tm <- transition_matrix(2, tmat, c(0.5, 0.5, 0, 1))
a <- health_state("Healthy", 1, 1, 0, FALSE)
b <- health_state("Dead", 1, 0, 0, TRUE)
health_states <- combine_state(a, b)
this.strategy <- strategy(tm, health_states, "intervention")
markov_model(this.strategy, 10, c(1, 0),c(0,0))
}
