% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_cost_analysis_functions.R
\name{microcosting_patches}
\alias{microcosting_patches}
\title{Function to estimate the cost of patches taken (from IPD)}
\usage{
microcosting_patches(ind_part_data, name_med, dose_med, dose_unit,
  no_taken, freq_taken, basis_time, unit_cost_data, unit_cost_column,
  cost_calculated_in, strength_expressed_in,
  list_period_timepoint = NULL, list_of_code_names = NULL,
  list_of_code_freq = NULL, list_of_code_dose_unit = NULL,
  equiv_dose = NULL)
}
\arguments{
\item{ind_part_data}{IPD}

\item{name_med}{name of the medication taken as patch}

\item{dose_med}{dose of medication}

\item{dose_unit}{unit of strength}

\item{no_taken}{how many taken}

\item{freq_taken}{frequency of medicaiton}

\item{basis_time}{basis for time}

\item{unit_cost_data}{unit costs data
assumes unit cost data contains the columns, name of medication, form/type, strength,
unit of strength, preparation, unitcost, size and size unit
(in which except name and form are to be passed in)}

\item{unit_cost_column}{column name of unit cost in unit_cost_data}

\item{cost_calculated_in}{column name of unit where the cost is calculated}

\item{strength_expressed_in}{column name of strength in the unit cost data}

\item{list_period_timepoint}{list of time period at each timepoint if they are coded}

\item{list_of_code_names}{if names are coded, give the code:name pairs}

\item{list_of_code_freq}{if frequency is coded, give the code:frequency pairs}

\item{list_of_code_dose_unit}{if unit is coded, give the code:unit pairs}

\item{equiv_dose}{if cost per equivalent doses are to be calcualted, provide equiv_dose}
}
\value{
the calculated cost of patches along with original data
}
\description{
Function to estimate the cost of patches taken (from IPD)
}
\examples{
med_costs_file = system.file("extdata", "average_unit_costs_med.csv", package = "packDAMipd")
data_file = system.file("extdata", "resource_use_p.csv", package = "packDAMipd")
ind_part_data = load_trial_data(data_file)
med_costs = load_trial_data(med_costs_file)
res=microcosting_patches(ind_part_data,"Drug", "patch_strength", "patch_dose_unit","patch_no_taken",
           "patch_frequency",  "day", med_costs, "UnitCost", "StrengthUnit","Strength",
           list(c("4 weeks", "1 week"), c(1,2)), NULL, NULL, NULL, "patch_equiv_dose")
}
