% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/4b_probabilistic_sensitivity_analysis_functions.R
\name{plot_ceac_psa}
\alias{plot_ceac_psa}
\title{Function to plot CEAC}
\usage{
plot_ceac_psa(
  control_markov,
  trt_markov,
  psa_table_ctrl,
  psa_table_trt,
  thresholds,
  num_rep,
  comparator
)
}
\arguments{
\item{control_markov}{markov model for control}

\item{trt_markov}{markov model for treatment}

\item{psa_table_ctrl}{param table for psa for control}

\item{psa_table_trt}{param table for psa for  treatment}

\item{thresholds}{threshold values of WTP}

\item{num_rep}{number of repetitions}

\item{comparator}{the strategy to be compared with}
}
\value{
plot of  ceac
}
\description{
Function to plot CEAC
}
