% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/4a_deterministic_sensitivity_analysis_functions.R
\name{plot_dsa_icer_range}
\alias{plot_dsa_icer_range}
\title{Function to do some checks before plotting sensitivity analysis results}
\usage{
plot_dsa_icer_range(ob_results, plot_var)
}
\arguments{
\item{ob_results}{results from deterministic sensitivity analysis}

\item{plot_var}{the variable}
}
\value{
plot
}
\description{
Function to do some checks before plotting sensitivity analysis results
}
