% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help_functions.R
\name{plot_meanSE_longitudinal_twogroups}
\alias{plot_meanSE_longitudinal_twogroups}
\title{Function to plot mean and SE for longitudinal observations for twogroups compared}
\usage{
plot_meanSE_longitudinal_twogroups(
  thedata,
  columnnames,
  timepoints,
  observation
)
}
\arguments{
\item{thedata}{the data where the observations are held}

\item{columnnames}{columnnames in the data where the intersted observations}

\item{timepoints}{the timepoints at which the descriptive analysis is done}

\item{observation}{name of the observations}
}
\value{
the plot that shows mean and SE
}
\description{
Function to plot mean and SE for longitudinal observations for twogroups compared
}
\examples{
eg_data <- as.data.frame(list(no = c(1, 2, 3, 4),
mark_at_1 = c(12, 7, 23, 45), gender = c("M", "F", "M", "F"),
mark_at_2 = c(12, 34, 89, 45), trialarm = c("1","1","2","2")))
plot_meanSE_longitudinal_twogroups(eg_data,  c("mark_at_1", "mark_at_2"),
c("1","2"), "mark")
}
