% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_parameter_estimation_survival.R
\name{plot_return_survival_curve}
\alias{plot_return_survival_curve}
\title{Plotting survival function for all covariates using survfit}
\usage{
plot_return_survival_curve(
  param_to_be_estimated,
  dataset,
  indep_var,
  covariates,
  timevar_survival
)
}
\arguments{
\item{param_to_be_estimated}{parameter to be estimated}

\item{dataset}{param describing the methods}

\item{indep_var}{independent variable}

\item{covariates}{covariates}

\item{timevar_survival}{time variable for survival analysis}
}
\value{
plot and the survival function values
}
\description{
Plotting survival function for all covariates using survfit
}
\examples{
\donttest{
 data_for_survival <- survival::lung
 plot_return_survival_curve(param_to_be_estimated = "status",
 dataset = data_for_survival,indep_var = "sex",covariates = c("ph.ecog"),
 timevar_survival = "time")
 }
}
