% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{promis3a_scoring.df}
\alias{promis3a_scoring.df}
\title{promis 3a scoring table}
\format{
A 14 by 3 dataframe
}
\source{
created on April 08, 2021
}
\usage{
promis3a_scoring.df
}
\description{
promis 3a scoring table
}
\keyword{datasets}
