% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_cost_analysis_functions.R
\name{return0_if_not_null_na}
\alias{return0_if_not_null_na}
\title{Function to return 0 if the param is not null or NA
trimming the white spaces}
\usage{
return0_if_not_null_na(param)
}
\arguments{
\item{param}{the form of medication either tablet or patch}
}
\value{
zero or -1
}
\description{
Function to return 0 if the param is not null or NA
trimming the white spaces
}
\examples{
parame = NULL
ans <- return0_if_not_null_na(parame)
parame = 1
ans <- return0_if_not_null_na(parame)
}
