% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_cost_analysis_functions.R
\name{return_equal_liststring_listcol}
\alias{return_equal_liststring_listcol}
\title{Function to get the subset of data compared to a string after
trimming the white spaces}
\usage{
return_equal_liststring_listcol(col, the_data, list_str)
}
\arguments{
\item{col}{the form of medication either tablet or patch}

\item{the_data}{the data to be get the subset from}

\item{list_str}{list of strings to be compared}
}
\value{
the subset data
}
\description{
Function to get the subset of data compared to a string after
trimming the white spaces
}
\examples{
the_data <- as.data.frame(cbind(c("one", "two"), c("tablet", "tablets"),
c("aa", "bb")))
colnames(the_data) <- c("name", "brand_a", "xx")
ans <- return_equal_liststring_listcol(2, the_data, c("tablet", "tablets"))
}
