% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_cost_analysis_functions.R
\name{return_equal_str_col}
\alias{return_equal_str_col}
\title{Function to get the subset of data compared to a string after
trimming the white spaces}
\usage{
return_equal_str_col(col, the_data, the_str)
}
\arguments{
\item{col}{the form of medication either tablet or patch}

\item{the_data}{the data to be get the subset from}

\item{the_str}{the string to be compared}
}
\value{
the subset data
}
\description{
Function to get the subset of data compared to a string after
trimming the white spaces
}
\examples{
the_data <- as.data.frame(cbind(c("one", "two"), c("a", "b")))
colnames(the_data) <- c("name", "brand")
ans <- return_equal_str_col(2, the_data, "a")
}
