% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_param}
\alias{table_param}
\title{Parameter table created}
\format{
A 11 by 2 dataframe
}
\source{
created on Jan 15, 2020
}
\usage{
table_param
}
\description{
Parameter table created
}
\keyword{datasets}
