% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trial_data}
\alias{trial_data}
\title{Example trial data}
\format{
A 31 by 33 dataframe
}
\source{
created on Jan 15, 2020
}
\usage{
trial_data
}
\description{
Example trial data
}
\keyword{datasets}
