% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_parameter_estimation_functions.R
\name{use_fh_survival}
\alias{use_fh_survival}
\title{##########################################################################################################
Get the parameter values using the survival analysis methid FH}
\usage{
use_fh_survival(param_to_be_estimated, dataset, method, indep_var,
  covariates_list, timevar_survival)
}
\arguments{
\item{param_to_be_estimated}{parameter of interest}

\item{dataset}{data set to be provided}

\item{method}{methd of estimation (for example, linear, logistic regression etc)}

\item{indep_var}{the independent variable (column name in data file)}

\item{covariates_list}{list of covariates - calculations to be done before passing}

\item{timevar_survival}{time variable for survival analysis, default is NA}
}
\value{
the results of the regression analysis
}
\description{
##########################################################################################################
Get the parameter values using the survival analysis methid FH
}
\examples{
data_for_survival<-survival::aml
surv_estimated_aml<-use_fh_survival("status", data_for_survival,
"survival analysis", "x", covariates_list=NA, "time")
}
