% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_parameter_estimation_functions.R
\name{use_km_survival}
\alias{use_km_survival}
\title{##########################################################################################################
Get the parameter values using the Kaplan-Meier survival analysis}
\usage{
use_km_survival(param_to_be_estimated, dataset, indep_var, covariates_list,
  timevar_survival)
}
\arguments{
\item{param_to_be_estimated}{parameter of interest}

\item{dataset}{data set to be provided}

\item{indep_var}{the independent variable (column name in data file)}

\item{covariates_list}{list of covariates - calculations to be done before passing}

\item{timevar_survival}{time variable for survival analysis, default is NA}
}
\value{
the results of the regression analysis
}
\description{
##########################################################################################################
Get the parameter values using the Kaplan-Meier survival analysis
}
\details{
This function is for regression using KM survival
}
\examples{
data_for_survival <- survival::aml
surv_estimated_aml <- use_km_survival("status", data_for_survival, "x",
covariates_list = NA, "time")
}
