% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_parameter_estimation_functions.R
\name{use_linear_rgression}
\alias{use_linear_rgression}
\title{##########################################################################################################
Get the parameter values using the linear regression}
\usage{
use_linear_rgression(param_to_be_estimated, dataset, indep_var, covariates,
  interaction = FALSE)
}
\arguments{
\item{param_to_be_estimated}{parameter of interest}

\item{dataset}{data set to be provided}

\item{indep_var}{the independent variable (column name in data file)}

\item{covariates}{list of covariates - calculations to be done before passing}

\item{interaction}{boolean value to indicate interaction in the case of linear regression,
false by default}

\item{method}{methd of estimation (for example, linear, logistic regression etc)}
}
\value{
the results of the regression analysis
}
\description{
##########################################################################################################
Get the parameter values using the linear regression
}
\examples{
mydata <- read.csv("https://stats.idre.ucla.edu/stat/data/binary.csv")
}
