% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_parameter_estimation_functions.R
\name{use_mixed_effect_model}
\alias{use_mixed_effect_model}
\title{Function for mixed effect regression}
\usage{
use_mixed_effect_model(param_to_be_estimated, dataset, indep_var,
  covariates = NA, random_effect = NA)
}
\arguments{
\item{param_to_be_estimated}{column name of dependent variable}

\item{dataset}{a dataframe}

\item{indep_var}{column name of independent variable}

\item{covariates, }{independent variables, NA by default}

\item{random_effect}{random effect variable(s) for the mixed effect models}
}
\value{
result regression result with plot if success and -1, if failure
}
\description{
Function for mixed effect regression
}
\examples{
mydata <- read.csv("https://stats.idre.ucla.edu/stat/data/binary.csv")
}
