% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_parameter_estimation_functions.R
\name{use_parametric_regression}
\alias{use_parametric_regression}
\title{##########################################################################################################
Get the parameter values using the survival analysis parametric regression}
\usage{
use_parametric_regression(param_to_be_estimated, dataset, indep_var,
  info_distribution, covariates_list, timevar_survival)
}
\arguments{
\item{param_to_be_estimated}{parameter of interest}

\item{dataset}{data set to be provided}

\item{indep_var}{the independent variable (column name in data file)}

\item{info_distribution}{distribution name  eg. for logistic regression -binomial}

\item{covariates_list}{list of covariates - calculations to be done before passing}

\item{timevar_survival}{time variable for survival analysis, default is NA}
}
\value{
the results of the regression analysis
}
\description{
##########################################################################################################
Get the parameter values using the survival analysis parametric regression
}
\examples{
data_for_survival<-survival::aml
surv_estimated_aml <- use_parametric_regression("status", data_for_survival,"x",
info_distribution="weibull", covariates_list=NA,"time")
}
