% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_parameter_estimation_functions.R
\name{use_parametric_survival}
\alias{use_parametric_survival}
\title{##########################################################################################################
Get the parameter values using the survival analysis parametric survival}
\usage{
use_parametric_survival(param_to_be_estimated, dataset, indep_var,
  info_distribution, covariates_list, timevar_survival)
}
\arguments{
\item{param_to_be_estimated}{parameter of interest}

\item{dataset}{data set to be provided}

\item{indep_var}{the independent variable (column name in data file)}

\item{info_distribution}{distribution name  eg. for logistic regression -binomial}

\item{covariates_list}{list of covariates - calculations to be done before passing}

\item{timevar_survival}{time variable for survival analysis, default is NA}
}
\value{
the results of the regression analysis
}
\description{
##########################################################################################################
Get the parameter values using the survival analysis parametric survival
}
\details{
This function is the last in the layer of function for parametric survival analysis
This then returns the parameteres of interest, plots the results etc
if the distribution is weibull it uses the package SurvRegCensCov for easy interpretation
of results
Returns the fit result, summary of regression, variance-covariance matrix of coeff,
cholesky decomposition, the parameters that define the assumed distrbution
and the plot of model prediction
check the basic assumption,
}
\examples{
data_for_survival <- survival::aml
surv_estimated_aml <- use_parametric_survival("status", data_for_survival, "x",
  info_distribution = "weibull", covariates_list = NA, "time"
)
}
