% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2c_outcome_cost_analysis_functions.R
\name{value_ADL_scores_IPD}
\alias{value_ADL_scores_IPD}
\title{Function to convert ADL scores to a T score}
\usage{
value_ADL_scores_IPD(ind_part_data, adl_related_words, adl_scoring,
  adl_scoring_data_columns, adl_nrcode)
}
\arguments{
\item{ind_part_data}{a dataframe containing IPD data}

\item{adl_related_words}{reltaed words to find out which columns contain adl data}

\item{adl_scoring}{ADL scoring table}

\item{adl_scoring_data_columns, }{ADL scoring table columnnames}

\item{adl_nrcode}{non response code for ADL}
}
\value{
ADLscores converted to T score included modified data, if success -1, if failure
}
\description{
Function to convert ADL scores to a T score based on
file:///C:/Users/smk543/Downloads/PROMIS%20Pain%20Interference%20Scoring%20Manual.pdf
}
\examples{
datafile = system.file("extdata", "trial_data.csv", package = "DecisionAnalysisModel")
trial_data = load_trial_data(datafile)
}
