% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2c_outcome_cost_analysis_functions.R
\name{value_Shows_IPD}
\alias{value_Shows_IPD}
\title{Function to convert Shows scores}
\usage{
value_Shows_IPD(ind_part_data, shows_related_words, shows_nrcode)
}
\arguments{
\item{ind_part_data}{a dataframe containing IPD}

\item{shows_related_words}{a dataframe containing IPD}

\item{shows_nrcode}{non response code for ADL, default is NA}
}
\value{
sum of scores, if success -1, if failure
}
\description{
Function to convert Shows scores
}
\examples{
datafile = system.file("extdata", "trial_data.csv", package = "DecisionAnalysisModel")
trial_data = load_trial_data(datafile)
value_Shows_IPD(trial_data,"qsy",NA)
}
