% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_cost_analysis_functions.R
\name{word2num}
\alias{word2num}
\title{Function to check the variable null or NA}
\usage{
word2num(word)
}
\arguments{
\item{word}{word for the number}
}
\value{
return the number
}
\description{
Function to check the variable null or NA
}
\details{
https://stackoverflow.com/questions/18332463/convert-written-number-to-number-in-r
}
\examples{
answer <- word2num("one forty one")
answer <- word2num("forty one and five hundred")
answer <- word2num("five thousand two hundred and eight")
}
