% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_help_functions.R
\name{add_probabilities}
\alias{add_probabilities}
\title{Function to add the probabilities in a vector}
\usage{
add_probabilities(probs)
}
\arguments{
\item{probs}{probabilities}
}
\value{
sum of probabilities if sucess else -1
}
\description{
Function to add the probabilities in a vector
}
\examples{
add_probabilities(c(0.6,0.2))
}
