% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_parameter_estimation_functions.R
\name{check_estimate_substitute_proper_params}
\alias{check_estimate_substitute_proper_params}
\title{#######################################################################
Function to get the expression, analyse it for what distribution it is,
then check if it has all parameters requires,
if not estimate the required parameters from given parameters,
and substitute it for actual expression that can
be used with R stats package}
\usage{
check_estimate_substitute_proper_params(the_expr)
}
\arguments{
\item{the_expr}{given parameters for generating ransom numbers}
}
\value{
the keyword that should be used in R for generating random numbers
}
\description{
#######################################################################
Function to get the expression, analyse it for what distribution it is,
then check if it has all parameters requires,
if not estimate the required parameters from given parameters,
and substitute it for actual expression that can
be used with R stats package
}
\details{
Get the expression of probability distribution with parameters, get the
parameters that are found, then find the required params and if they are ok
if not try if they can be estimated. Then it will substitute for proper
parameters in the way R is expecting to perform
}
\examples{
check_estimate_substitute_proper_params("gamma(mean = 10 ,sd=1)")
}
\keyword{internal}
