% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_model_functions.R
\name{define_parameters}
\alias{define_parameters}
\title{Function to return a list of parameters given}
\usage{
define_parameters(...)
}
\arguments{
\item{...}{any parameters set of name value pairs expected}
}
\value{
a list of parameters
}
\description{
Function to return a list of parameters given
}
\details{
To return a list of parameters
For using with assign_parameters() just list or enumerate the
parameters, do not use c() or list() to create a data type list
}
\examples{
define_parameters(rr = 1)
}
