% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_model_functions.R
\name{define_transition_table}
\alias{define_transition_table}
\title{Define the table for transition}
\usage{
define_transition_table(tmat)
}
\arguments{
\item{tmat}{transition matrix in the format as in package 'mstate'}
}
\value{
the transition table with the probabilities
}
\description{
Define the table for transition
}
\details{
Generating a table for transition matrix for efficient understanding and
checking
The transition matrix in the format as per 'mstate' package is
transformed to a table. if tmat is not a square matrix, it gives error
else it spells out the transition number, probability name and from state
to state
}
\examples{
tmat <- rbind(c(1, 2), c(3, 4))
colnames(tmat) <- rownames(tmat) <- c("Healthy", "Dead")
define_transition_table(tmat)
}
