% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_parameter_estimation_functions.R
\name{find_glm_distribution}
\alias{find_glm_distribution}
\title{Function to find the keyword for family of distribution in glm}
\usage{
find_glm_distribution(text)
}
\arguments{
\item{text}{distribution}
}
\value{
the keyword - the name of distribution
}
\description{
Function to find the keyword for family of distribution in glm
}
\details{
Find the family for glm method
}
\examples{
find_glm_distribution("gamma")
}
