% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_parameter_estimation_functions.R
\name{find_required_parameter_combs}
\alias{find_required_parameter_combs}
\title{Function to find the parameters that determine the probability distribution}
\usage{
find_required_parameter_combs(name_distri)
}
\arguments{
\item{name_distri}{name of the probability distribution}
}
\value{
the parameters that determine the distribution
}
\description{
Function to find the parameters that determine the probability distribution
}
\details{
For each of the probability distribution we require certain parameters
and this function provides that required list of parameters.
}
\examples{
find_required_parameter_combs("gamma")
}
