% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_parameter_estimation_functions.R
\name{form_expression_lm}
\alias{form_expression_lm}
\title{Form expression to use with lm()}
\usage{
form_expression_lm(param_to_be_estimated, indep_var, covariates, interaction)
}
\arguments{
\item{param_to_be_estimated}{parameter of interest}

\item{indep_var}{the independent variable (column name in data file)}

\item{covariates}{list of covariates}

\item{interaction}{boolean value to indicate interaction
in the case of linear regression, false by default}
}
\value{
the formula for lm
}
\description{
Form expression to use with lm()
}
\details{
This function helps to create the expression for liner regression model
it takes care of covariates and interaction
}
\examples{
formula <- form_expression_lm("gre", indep_var = "gpa", covariates = NA,
interaction = FALSE)
}
