% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3a_trialdata_analysis_input_functions.R
\name{get_age_details}
\alias{get_age_details}
\title{Function to get the details of the age column}
\usage{
get_age_details(trialdata)
}
\arguments{
\item{trialdata, }{data containing individual level trial data}
}
\value{
the name of the variable related to age and the unique contents
if success, else error
}
\description{
Function to get the details of the age column
}
\details{
expecting the data contains the information on age
preferably  column names "age", "dob" or "yob" or "date of birth".
"year of birth", "birth year"
If multiple column names match these, then first match will be chosen.
}
\examples{
get_age_details(data.frame("Age" = c(21, 15),
"arm" = c("control", "intervention")))
}
