% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_parameter_estimation_functions.R
\name{get_extension_file}
\alias{get_extension_file}
\title{Function to get extension of a file name}
\usage{
get_extension_file(filename)
}
\arguments{
\item{filename}{name of a file}
}
\value{
the extension
}
\description{
Function to get extension of a file name
}
\details{
if there is no "." character returns error
else returns last characters those after string split using "."
}
\examples{
get_extension_file("data.txt")
}
