% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_parameter_estimation_functions.R
\name{get_mortality_from_file}
\alias{get_mortality_from_file}
\title{Get the parameter values from reading a file}
\usage{
get_mortality_from_file(paramfile, age, gender = NULL)
}
\arguments{
\item{paramfile}{parameter file to get the mortality eg.national life table data}

\item{age}{age to get the age specific data}

\item{gender}{gender details to get the gender specific mortality data}
}
\value{
the paramvalue
}
\description{
Get the parameter values from reading a file
}
\examples{
paramfile <- system.file("extdata", "LifeTable_USA_Mx_2015.csv",
  package = "packDAMipd"
)
a <- get_mortality_from_file(paramfile, age = 10, gender = NULL)
}
