% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_parameter_estimation_functions.R
\name{get_parameter_direct}
\alias{get_parameter_direct}
\title{Get the parameter values from reading a file}
\usage{
get_parameter_direct(parameter, paramvalue)
}
\arguments{
\item{parameter}{parameter of interest}

\item{paramvalue}{parameter value to be assigned}
}
\value{
the paramvalue
}
\description{
Get the parameter values from reading a file
}
\details{
Basic function to assign a parameter directly
}
\examples{
a <- get_parameter_direct("cost_IT", paramvalue = 100)
}
