% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_parameter_estimation_functions.R
\name{get_parameter_estimated_regression}
\alias{get_parameter_estimated_regression}
\title{Get the parameter values using the provided statistical regression methods}
\usage{
get_parameter_estimated_regression(param_to_be_estimated, dataset, method,
  indep_var, info_get_method, info_distribution, covariates = NA,
  strategycol = NA, strategyname = NA, timevar_survival = NA,
  interaction = FALSE)
}
\arguments{
\item{param_to_be_estimated}{parameter of interest}

\item{dataset}{data set to be provided}

\item{method}{methd of estimation (for example, linear, logistic regression etc)}

\item{indep_var}{the independent variable (column name in data file)}

\item{info_get_method}{additional information on methods e.g Kaplan-Meier ot hazard}

\item{info_distribution}{distribution name  eg. for logistic regression -binomial}

\item{covariates}{list of covariates - calculations to be done before passing}

\item{strategycol}{column name containing arm details, default is NA}

\item{strategyname}{name of the  arm, default is NA}

\item{timevar_survival}{time variable for survival analysis, default is NA}

\item{interaction}{boolean value to indicate interaction in the case of linear regression,
false by default}
}
\value{
the results of the regression analysis
}
\description{
Get the parameter values using the provided statistical regression methods
}
\examples{
mydata <- read.csv("https://stats.idre.ucla.edu/stat/data/binary.csv")
results_logit <- get_parameter_estimated_regression("admit", mydata,
"logistic regression", "gre", NA,"binomial", c("gpa", "factor(rank)"))
}
