% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_parameter_estimation_functions.R
\name{load_trial_data}
\alias{load_trial_data}
\title{Function to load the file containing trial data and return it}
\usage{
load_trial_data(file = NULL, sheet = NULL)
}
\arguments{
\item{file, }{name of the file in full}

\item{sheet}{name of the sheet if excel work book is given}
}
\value{
trial data if success, else -1
}
\description{
Function to load the file containing trial data and return it
}
\examples{
load_trial_data(system.file("extdata", "trial_data.csv",
  package = "packDAMipd"
))
}
