% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_parameter_estimation_functions.R
\name{use_logistic_rgression}
\alias{use_logistic_rgression}
\title{##########################################################################################################
Get the parameter values using logistic regression}
\usage{
use_logistic_rgression(param_to_be_estimated, dataset, indep_var, family,
  covariates_list, naaction, link)
}
\arguments{
\item{param_to_be_estimated}{parameter of interest}

\item{dataset}{data set to be provided}

\item{indep_var}{the independent variable (column name in data file)}

\item{family}{distribution name  eg. for logistic regression -binomial}

\item{covariates_list}{list of covariates - calculations to be done before passing}

\item{naaction}{action to be taken with the missing values}

\item{link}{link function if not the default for each family}
}
\value{
the results of the regression analysis
}
\description{
##########################################################################################################
Get the parameter values using logistic regression
}
\examples{
mydata <- read.csv("https://stats.idre.ucla.edu/stat/data/binary.csv")
mydata$rank <- factor(mydata$rank)
results_logit <- use_logistic_rgression("admit", dataset=mydata,
indep_var = "gre", family ="binomial", covariates_list = NA, naaction = "stats::na.omit", link = NA)
}
